+++ frontend/pages/store.js
@@ -1,8 +1,8 @@
 import Link from "next/link";
-import fetch from "isomorphic-unfetch";
 
 import Layout from "../components/Layout";
 import Footer from "../components/Footer";
+import { fetchData } from "../utils/fetchData";
 
 function Store({
   products,
@@ -37,11 +37,11 @@ function Store({
                 className="fl w-100 w-50-m w-33-l pa2"
                 key={product.id}
               >
-                <Link href={`/model?id=${product.id}`}>
+                <Link href={`/model/${product.id}`}>
                   <a className="db link dim tc white">
                     <img
                       style={{ objectFit: "contain", height: "200px" }}
-                      src={`/static/models/${product.id}/thumbnail@m.jpg`}
+                      src={`/models/${product.id}/thumbnail@m.jpg`}
                       alt="Lorem"
                       className="w-100 db outline black-10"
                     />
@@ -66,29 +66,14 @@ function Store({
   );
 }
 
-Store.getInitialProps = async ({ req }) => {
-  let URL;
+export async function getStaticProps() {
+  const products = await fetchData(`/api/get-products`);
 
-  if (typeof window === "undefined") {
-    if (process.env.NODE_ENV === "production") {
-      URL = `https://${req.headers.host}/api/get-products`;
-    } else {
-      URL = `https://localhost:3000/api/get-products`;
-    }
-  } else {
-    URL = "/api/get-products";
-  }
-
-  const props = { products: [] };
-
-  try {
-    const response = await fetch(URL);
-    const { docs } = await response.json();
-    props.products = docs;
-  } catch (e) {
-    console.error(e.message);
-  }
-  return props;
+  return {
+    props: {
+      products
+    },
+  };
 };
 
 export default Store;
